Fix number display in details page

+++ users/models.py
@@ -47,7 +47,7 @@ class PointsType(models.Model):
 class Point(models.Model):
     user = models.ForeignKey(CustomUser, on_delete=models.CASCADE, related_name='point', null=True)
     type = models.ForeignKey(PointsType, on_delete=models.CASCADE, related_name='type', null=True)
-    value = models.IntegerField()
+    value = models.FloatField()
     details = models.CharField(max_length=256, default='')
     record_date = models.DateField(default=now)
